Bugfix for PlayingMessage Deletion
This commit makes a small bugfix that ensures a message isn't
deleted for a second time when the bot drops out of a VC channel.
This may introduce a bug where the message is left if the bot is
manually disconnected from the VC. Also adds a helpful comment
for registering commans globally.