Partial Implementation of Reaction Roles
This commit adds a slash command /role with 2 subcommands, add and
reset. /role add allows the caller to mark an existing role as a
reaction role and provide the emoji and optional description for the
reaction role message. /role reset allows the caller to mark an existing
message as a reaction role message, and the bot will fill up its
reaction role list using the previous messsage's content and use that
message for future roles. Essentially, this command covers the standard
usage for CREATING reaction role messages.

However, no functionality for listening to reaction updates and
assigning roles is implemented here. Adding reaction updates to the
bot's intents will significantly increase the traffic to the bot, as
there is no way to filter for just reactions on the relevant message
server-side. Since this bot is running on a VM with limited network
capacity, I'd prefer an alternate method of handling this functionality.
Pushing this commit onto a feature branch while I consider.