Adds Detailed Help Text Option
This commit adds a `longHelp` field to the Command type which contains
long-form help text for more complex commands. It also renames the
existing `help` field to `shortHelp`, since it is explicitly short-form
help text now. It also extends the `help` SlashCommand with an optional
string argument called `command` that takes a command name (no
autocomplete unfortunately). When that command has a long-form help text
provided (not required), the bot sends it, otherwise it sends an
ephemeral error message directing users to the normal help summary.